# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: All-Aboard API
  description: |
    Contrat HTTP Phase 2 MVP (feed, auth JWT, demandes d'aide, mentor feed).
    Source de vérité types : `packages/types`. Doc opérationnelle : `Docs/guides/web-api-integration.md`.
  version: 0.10.0
  contact:
    name: All-Aboard THP
    url: https://github.com/AllAboard-THP/All-Aboard
servers:
  - url: http://localhost:4000
    description: Développement local
  - url: https://api-staging.allaboard.fr
    description: Staging Dokploy
...
paths:
...
  /auth/login:
    post:
      operationId: postAuthLogin
      tags: [auth]
      summary: Connexion MVP (mot de passe partagé)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: JWT émis (cookie httpOnly `access_token` + claims sub/role)
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Cookie `access_token` (httpOnly, SameSite=Lax)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Corps invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid_body
        "401":
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid_credentials
        "503":
          description: MVP_LOGIN_PASSWORD non configuré
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: login_not_configured
  /auth/register:
    post:
      operationId: postAuthRegister
      tags: [auth]
      summary: Inscription (hash argon2, ADR 0003)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "200":
          description: Compte créé + JWT (cookie)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Corps invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email déjà utilisé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: email_taken
        "503":
          description: Base indisponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Utilisateur courant (JWT + profil DB)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profil enrichi si DB disponible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "401":
          description: JWT absent ou invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT émis par POST /auth/login (claim sub = userId)
...
  schemas:
...
    UserRole:
      type: string
      enum: [student, mentor, admin]
...
    SubjectSummary:
      type: object
      required: [id, name, slug, icon, accentColor]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        accentColor:
          type: string
...
    LoginBody:
      type: object
      required: [password]
      properties:
        email:
          type: string
          format: email
          maxLength: 256
        userId:
          type: string
          description: Alias legacy dev (bob, alice) — ignoré en staging/prod
          minLength: 1
          maxLength: 256
        password:
          type: string
          minLength: 1
          maxLength: 256
      oneOf:
        - required: [email]
        - required: [userId]
    LoginResponse:
      type: object
      required: [ok, userId, role]
      properties:
        ok:
          type: boolean
          const: true
        userId:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
    AuthMeResponse:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        role:
          $ref: "#/components/schemas/UserRole"
        displayName:
          type: string
        fullName:
          type: string
        headline:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        educationLevel:
          type: string
        cguAcceptedAt:
          type: string
          format: date-time
        notifyOnComment:
          type: boolean
        notifyOnMessage:
          type: boolean
        certificationTags:
          type: array
          items:
            type: string
        competenceSubjects:
          type: array
          items:
            $ref: "#/components/schemas/SubjectSummary"
    RegisterBody:
      type: object
      required: [email, password, fullName, acceptCgu]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        passwordConfirmation:
          type: string
          minLength: 8
        fullName:
          type: string
          minLength: 1
        educationLevel:
          type: string
        headline:
          type: string
        acceptCgu:
          type: boolean
          const: true
    RegisterResponse:
      type: object
      required: [ok, userId, role]
      properties:
        ok:
          type: boolean
          const: true
        userId:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...